+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./baseline.js";
 export * from "./compare.js";
 export * from "./config.js";
+export * from "./custom-rules.js";
 export * from "./data-path.js";
 export * from "./diagnostics.js";
 export * from "./errors.js";